requirements.txt
aiogram==3.18.0

# database
SQLAlchemy==2.0.38
asyncpg==0.30.0
alembic==1.16.5

# ИИ
openai==1.63.0

# env
python-dotenv==1.0.1

# log
colorlog==6.9.0

# timezone
pytz==2025.1